[Changed] Add/init webSockPort & httpJSONPort in net/node By the way, fix httpInfoPort from int to uint16

+++ net/protocol/protocol.go
@@ -30,7 +30,7 @@ type Noder interface {
 	Services() uint64
 	GetAddr() string
 	GetPort() uint16
-	GetHttpInfoPort() int
+	GetHttpInfoPort() uint16
 	SetHttpInfoPort(uint16)
 	GetHttpInfoState() bool
 	SetHttpInfoState(bool)
